server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
llama-cpp-python==0.2.90
python-dotenv==1.0.1
starlette==0.41.2
